requirements.txt
aiosqlite==0.20.0
alembic==1.13.1
annotated-types==0.6.0
anyio==3.7.1
astroid==3.2.2
async-timeout==4.0.3
asyncpg==0.29.0
cachetools==5.4.0
certifi==2024.2.2
cffi==1.16.0
charset-normalizer==3.3.2
click==8.1.7
coverage==7.6.1
cryptography==41.0.7
dill==0.3.8
dnspython==2.5.0
ecdsa==0.18.0
email-validator==2.1.0.post1
exceptiongroup==1.2.0
fastapi==0.104.1
google-api-core==2.19.1
google-auth==2.32.0
google-auth-httplib2==0.2.0
googleapis-common-protos==1.63.2
greenlet==3.0.1
h11==0.14.0
httpcore==1.0.5
httplib2==0.22.0
httpx==0.27.0
idna==3.6
iniconfig==2.0.0
isort==5.13.2
Mako==1.3.3
MarkupSafe==2.1.5
mccabe==0.7.0
oauthlib==3.2.2
packaging==24.0
passlib==1.7.4
pillow==10.4.0
platformdirs==4.2.2
pluggy==1.4.0
proto-plus==1.24.0
protobuf==5.27.3
psycopg2-binary==2.9.9
pyasn1==0.5.1
pyasn1_modules==0.4.0
pycparser==2.21
pydantic==2.8.2
pydantic-settings==2.1.0
pydantic_core==2.20.1
pyparsing==3.1.2
pytest==8.1.1
pytest-cov==5.0.0
pytest-mock==3.14.0
python-dotenv==1.0.0
python-jose==3.3.0
python-multipart==0.0.6
pytz==2023.3.post1
requests==2.32.3
requests-oauthlib==2.0.0
rsa==4.9
six==1.16.0
sniffio==1.3.0
SQLAlchemy==2.0.23
SQLAlchemy-Utils==0.41.2
starlette==0.27.0
tomli==2.0.1
tomlkit==0.12.5
typing_extensions==4.8.0
uritemplate==4.1.1
urllib3==2.2.2
uvicorn==0.24.0.post1
